Re-add previously removed episodes (#62)
* Change recommended autocrlf config for Windows (#60)

Changes the Windows config recommendation to true, aligning it with the MacOS/Linux.

* Add back in some removed episodes

Also return `undoing changes` and `ignoring files` to their original order. Listed timings add up to more than is suitable for the schedule (needs to be updated for the May course). Did not add back in episodes 17 (github issues) and 18 (remote branches from git).